Rename DriverSlowTrigger to AxisThresholdTrigger, add hand to contructor, add more tests
Further generalize by renaming and letting the user set the hand to use, and add a test for if the trigger is close to the threshold or right on it.
