+++ src/main/java/com/systemmeltdown/robotlib/triggers/AxisThresholdTrigger.java
@@ -4,22 +4,20 @@ import edu.wpi.first.wpilibj.XboxController;
 import edu.wpi.first.wpilibj.GenericHID.Hand;
 import edu.wpi.first.wpilibj.buttons.Trigger;
 
-public class DriverSlowTrigger extends Trigger {
+public class AxisThresholdTrigger extends Trigger {
   private XboxController controller;
+  private Hand hand;
   private double triggerThreshold;
 
-  public DriverSlowTrigger(XboxController controller) {
-    this(controller, 0.25);
-  }
-
-  public DriverSlowTrigger(XboxController controller, double triggerThreshold) {
+  public AxisThresholdTrigger(XboxController controller, Hand hand, double triggerThreshold) {
     this.controller = controller;
+    this.hand = hand;
     this.triggerThreshold = triggerThreshold;
   }
 
   @Override
   public boolean get() {
-    double axis = controller.getTriggerAxis(Hand.kRight);
+    double axis = controller.getTriggerAxis(hand);
 
     return (axis > triggerThreshold);
   }
